survey_v2
I've been working on some modifications to the customer feedback system created by Kavithvajen and thought I'd share what I've done in case anyone wants to make any suggestions on how it can be improved. I can't claim any credit for this as it's technically the first Pi based project that I've made. I simply took the supplied code, and then added a copy of the code for a third button used as the "ok" button, and then in a separate file I created a new script to either launch the main code or copy the database out to a file with the name being the time and date it was copied, and then deleting the database file before launching the main code. The only other change I made was to remove the emoji display when a button is pressed as it was holding the code up and if you have lots of people trying to give feedback then this could cause a bottleneck. There's probably a much better way to do this all in one file but I'm no coder, so this is just my way that I got it to work. Please feel free to offer updates and improvements. I'm also new to github, so I'll try to work out how to post the other file that I use to call the main script next.