Add files via upload
Here are the files for my updated version of this project. I'm sure that there is an easy way to combine this all into one file, but this is my first Pi project and I'm not a coder, so please feel free to let me know what (if anything) is wrong and how to improve.